chore(localfile): rename package tofile to localfile
