+++ pkg/controller/controller.go
@@ -9,7 +9,7 @@ import (
 	"github.com/suzuki-shunsuke/tfcmt/pkg/config"
 	"github.com/suzuki-shunsuke/tfcmt/pkg/notifier"
 	"github.com/suzuki-shunsuke/tfcmt/pkg/notifier/github"
-	"github.com/suzuki-shunsuke/tfcmt/pkg/notifier/tofile"
+	"github.com/suzuki-shunsuke/tfcmt/pkg/notifier/localfile"
 	tmpl "github.com/suzuki-shunsuke/tfcmt/pkg/template"
 	"github.com/suzuki-shunsuke/tfcmt/pkg/terraform"
 )
@@ -133,7 +133,7 @@ func (ctrl *Controller) getNotifier(ctx context.Context) (notifier.Notifier, err
 	}
 	// Write output to file instead of github comment
 	if ctrl.Config.Output != "" {
-		client, err := tofile.NewClient(ctx, &tofile.Config{
+		client, err := localfile.NewClient(ctx, &localfile.Config{
 			OutputFile:         ctrl.Config.Output,
 			Parser:             ctrl.Parser,
 			UseRawOutput:       ctrl.Config.Terraform.UseRawOutput,
