+++ pkg/notifier/localfile/client.go
@@ -1,4 +1,4 @@
-package tofile
+package localfile
 
 import (
 	"context"
