+++ pkg/notifier/localfile/notify.go
@@ -1,4 +1,4 @@
-package tofile
+package localfile
 
 // NotifyService handles communication with the notification related
 // methods of GitHub API
